WFP-1127 remove test duplication

+++ src/test/kotlin/uk/gov/justice/digital/hmpps/hmppsworkload/integration/IntegrationTestBase.kt
@@ -170,7 +170,7 @@ abstract class IntegrationTestBase {
     )
   }
 
-  protected fun staffIdResponse(staffId: Long, staffCode: String, teamCode: String, staffGradeCode: String = "PSM") {
+  protected fun staffIdResponse(staffId: BigInteger, staffCode: String, teamCode: String, staffGradeCode: String = "PSM") {
     val request = HttpRequest.request().withPath("/staff/staffIdentifier/$staffId")
     communityApi.`when`(request, Times.exactly(1)).respond(
       HttpResponse.response().withContentType(MediaType.APPLICATION_JSON).withBody(staffByIdResponse(staffCode, staffGradeCode, teamCode))
